Configurando botões de regra em todas as telas usando o controller
