+++ app/src/main/java/com/ifmg/termix/DailyGame.kt
@@ -36,10 +36,11 @@ class DailyGame : AppCompatActivity() {
             insets
         }
 
+        // Criar instância do controller
         gameController = GameController(this)
 
         // Registrar eventos
-        registerButtonEvents()
+        registerButtonEvents(gameController)
         createLettersGrid()
         createKeyBoardGrid()
 
@@ -168,7 +169,7 @@ class DailyGame : AppCompatActivity() {
     }
 
     // Configurar todos os eventos de botão
-    private fun registerButtonEvents(){
+    private fun registerButtonEvents(gameController: GameController){
 
         // Voltar à tela inicial
         dailyGameBinding.backToHomeDailyBtn.setOnClickListener {
@@ -180,6 +181,11 @@ class DailyGame : AppCompatActivity() {
         dailyGameBinding.profileDailyBtn.setOnClickListener {
             val intent = Intent(this, Profile::class.java)
             startActivity(intent)
+        }
+
+        // Mostrar regras do jogo em um pop up customizado
+        dailyGameBinding.ruleDailyBtn.setOnClickListener {
+            gameController.showPopup(R.layout.daily_rules)
         }
     }
