+++ app/src/main/java/com/ifmg/termix/InfiniteMinigame.kt
@@ -1,22 +1,20 @@
 package com.ifmg.termix
 
 import android.content.Intent
-import android.graphics.Color
-import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
 import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.ifmg.termix.controller.GameController
 import com.ifmg.termix.databinding.ActivityInfiniteMinigameBinding
 
 
 class InfiniteMinigame : AppCompatActivity() {
 
     private lateinit var infiniteMinigameBinding: ActivityInfiniteMinigameBinding
+
+    private lateinit var gameController: GameController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -33,12 +31,15 @@ class InfiniteMinigame : AppCompatActivity() {
             insets
         }
 
-        registerButtonEvents()
+        // Criar instância do controller
+        gameController = GameController(this)
 
+        // Registrar eventos
+        registerButtonEvents(gameController)
     }
 
     // Configurar todos os eventos de botão
-    private fun registerButtonEvents(){
+    private fun registerButtonEvents(gameController: GameController){
 
         // Voltar à tela de minijogos
         infiniteMinigameBinding.backToHomeInfiniteBtn.setOnClickListener {
@@ -52,22 +53,9 @@ class InfiniteMinigame : AppCompatActivity() {
             startActivity(intent)
         }
 
-        // TODO: adicionar em um método intermediário e criar evento do botão nos demais jogos
         // Mostrar regras do jogo em um pop up customizado
         infiniteMinigameBinding.ruleInfiniteBtn.setOnClickListener {
-            val alertCustomdialog: View = LayoutInflater.from(this@InfiniteMinigame).inflate(R.layout.infinite_rules, null)
-            val alert: AlertDialog.Builder = AlertDialog.Builder(this)
-            alert.setView(alertCustomdialog)
-
-            val dialog: AlertDialog = alert.create()
-            dialog.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
-            dialog.show()
-
-            // Fechar o pop up pode ser fechado ao clicar nele
-            alertCustomdialog.setOnClickListener {
-                dialog.dismiss()
-            }
-
+            gameController.showPopup(R.layout.infinite_rules)
         }
 
     }
